Add error handling to service classes

+++ server/src/main/java/service/UserService.java
@@ -2,6 +2,7 @@ package service;
 import dataaccess.*;
 import model.*;
 
+import java.util.Objects;
 import java.util.UUID;
 
 public class UserService {
@@ -16,36 +17,52 @@ public class UserService {
     }
 
     public RegisterResult register(RegisterRequest registerRequest) throws DataAccessException {
+        if ((registerRequest.username() == null)||(registerRequest.email()==null)||(registerRequest.password()==null)){
+            throw new DataAccessException("Error: bad request");
+        }
         UserData userData = userDAO.getUser(registerRequest.username());
-        if (userData == null){
+        if (userData != null){
+            throw new DataAccessException("Error: already taken");
+        }
+        try{
             UserData newData = new UserData(registerRequest.username(), registerRequest.password(), registerRequest.email());
             userDAO.insertUser(newData);
             String newToken = generateToken();
             AuthData newAuth = new AuthData(newToken, registerRequest.username());
             authDAO.createAuth(newAuth);
             return new RegisterResult(registerRequest.username(), newToken);
-        }
-        else{
-            throw new DataAccessException("username already taken");
+        } catch (DataAccessException e){
+            throw new DataAccessException("Error: %e");
         }
     }
 
     public LoginResult login(LoginRequest loginRequest) throws DataAccessException {
         UserData userData = userDAO.getUser(loginRequest.username());
-        if (userData != null){
+        if (userData == null){
+            throw new DataAccessException("Error: Username doesn't exist");
+        }
+        if (!Objects.equals(userData.password(), loginRequest.password())){
+            throw new DataAccessException("Error: unauthorized");
+        }
+        try{
             String newToken = generateToken();
             AuthData newAuth = new AuthData(newToken, loginRequest.username());
             authDAO.createAuth(newAuth);
             return new LoginResult(loginRequest.username(), newToken);
+        } catch (DataAccessException e){
+            throw new DataAccessException("Error: %e");
         }
-        else{
-            throw new DataAccessException("username doesn't exist");
-        }
-
     }
 
-    public void logout(LogoutRequest logoutRequest) {
+    public void logout(LogoutRequest logoutRequest) throws DataAccessException {
         AuthData authData = authDAO.getAuth(logoutRequest.authToken());
-        authDAO.deleteAuth(authData.authToken());
+        if (authData == null){
+            throw new DataAccessException("Error: unauthorized");
+        }
+        try{
+            authDAO.deleteAuth(authData.authToken());
+        } catch (DataAccessException e){
+            throw new DataAccessException("Error: %e");
+        }
     }
 }
